tentando remover recursão
