--- src/sintatico/temp.py
@@ -1,21 +1,0 @@
-producao_programa = Producao("PROGRAMA")
-producao_ld = Producao("LISTA DECLARAÇÃO")
-producao_ld1 = Producao("LISTA DECLARAÇÃO'")
-producao_decl = Producao("DECLARAÇÃO")
-producao_var_decl = Producao("VAR-DECLARAÇÃO")
-producao_esp_tipo = Producao("ESPECIFICADOR-TIPO")
-producao_fun_decl = Producao("FUN-DECLARAÇÃO")
-producao_params = Producao("PARAMETROS")
-producao_params_list = Producao("PARAMETRO-LIST")
-producao_params_list1 = Producao("PARAMETRO-LIST'")
-producao_param = Producao("PARAMETRO")
-producao_compound_stmt = Producao("COMPOUND-STMT")
-producao_decl_loc = Producao("DECLARACAO-LOCAIS")
-producao_decl_loc1 = Producao("DECLARACAO-LOCAIS'")
-producao_stmt_list = Producao("STATEMENT-LIST")
-producao_stmt_list1 = Producao("STATEMENT-LIST'")
-producao_exp_stmt = Producao("EXPRESSAO-STMT")
-producao_sel_stmt = Producao("SELECAO-STMT")
-producao_iter_stmt = Producao("ITERACAO-STMT")
-producao_ret_stmt = Producao("RETURN-STMT")
-producao_exp_stmt = Producao("EXPRESSAO")
